Update AnomalousCouplingEFTNegative.py
Changing the way the sm component yield is retrieved. If process contains "_sm" it will return "func_sm" but this is also true for "WW_sm_lin_quad_OP" for example. Checking the end of the string, if it ends with "sm" then we are sure this is the sm component